refactor: move reporting types to "compose/" folder
