+++ src/compose/reporting/types.ts
@@ -1,0 +1,5 @@
+export type TestResult = {
+    title: string;
+    status: "succeed" | "failed" | "error" | "not runed" | "canceled";
+    children?: TestResult[];
+};
