+++ src/index.ts
@@ -1,6 +1,5 @@
-export { mapScenarios } from "@compose/map-scenarios";
+export { mapScenarios, type TestResult } from "@compose";
 export { createScenario as scenario } from "@modules/scenarios";
 export { createReusableTest as reusableTest } from "@modules/reusable-tests";
 export { listeners } from "@modules/listeners";
 export { expect } from "@core/expectations";
-export { type TestResult } from "@shared/types";
